debt - merge goToDefinition and referenceSearch into gotoSymbol, extract peekView
