+++ src/vs/editor/contrib/gotoSymbol/goToSymbol.ts
