+++ src/vs/editor/contrib/gotoSymbol/link/clickLinkGesture.ts
