+++ src/vs/editor/contrib/gotoSymbol/link/goToDefinitionAtPosition.css
