+++ src/vs/editor/contrib/gotoSymbol/peek/referencesWidget.css
