+++ src/vs/editor/contrib/gotoSymbol/referencesModel.ts
