+++ src/vs/editor/contrib/gotoSymbol/test/referencesModel.test.ts
@@ -2,11 +2,12 @@
  *  Copyright (c) Microsoft Corporation. All rights reserved.
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
+
 import * as assert from 'assert';
 import { URI } from 'vs/base/common/uri';
 import { Range } from 'vs/editor/common/core/range';
 import { Position } from 'vs/editor/common/core/position';
-import { ReferencesModel } from 'vs/editor/contrib/referenceSearch/referencesModel';
+import { ReferencesModel } from 'vs/editor/contrib/gotoSymbol/referencesModel';
 
 suite('references', function () {
