+++ src/vs/editor/contrib/peekView/media/peekViewWidget.css
