+++ src/vs/editor/editor.all.ts
@@ -22,8 +22,8 @@ import 'vs/editor/contrib/find/findController';
 import 'vs/editor/contrib/folding/folding';
 import 'vs/editor/contrib/fontZoom/fontZoom';
 import 'vs/editor/contrib/format/formatActions';
-import 'vs/editor/contrib/goToDefinition/goToDefinitionCommands';
-import 'vs/editor/contrib/goToDefinition/goToDefinitionAtPosition';
+import 'vs/editor/contrib/gotoSymbol/goToCommands';
+import 'vs/editor/contrib/gotoSymbol/link/goToDefinitionAtPosition';
 import 'vs/editor/contrib/gotoError/gotoError';
 import 'vs/editor/contrib/hover/hover';
 import 'vs/editor/contrib/inPlaceReplace/inPlaceReplace';
@@ -31,7 +31,6 @@ import 'vs/editor/contrib/linesOperations/linesOperations';
 import 'vs/editor/contrib/links/links';
 import 'vs/editor/contrib/multicursor/multicursor';
 import 'vs/editor/contrib/parameterHints/parameterHints';
-import 'vs/editor/contrib/referenceSearch/referenceSearch';
 import 'vs/editor/contrib/rename/rename';
 import 'vs/editor/contrib/smartSelect/smartSelect';
 import 'vs/editor/contrib/snippet/snippetController2';
